Bugfix: repo's defined in @EnableJpaRepositories are not allowed to have the same base-package
