+++ core/src/main/java/com/ritense/valtimo/service/ChoiceFieldService.java
@@ -18,7 +18,7 @@ package com.ritense.valtimo.service;
 
 import com.ritense.valtimo.domain.choicefield.ChoiceField;
 import com.ritense.valtimo.domain.choicefield.ChoiceFieldValue;
-import com.ritense.valtimo.repository.ChoiceFieldRepository;
+import com.ritense.valtimo.choicefield.repository.ChoiceFieldRepository;
 import com.ritense.valtimo.web.rest.dto.ChoiceFieldDTO;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
