+++ core/src/main/java/com/ritense/valtimo/service/ContextService.java
@@ -21,8 +21,8 @@ import com.ritense.valtimo.contract.authentication.model.ValtimoUser;
 import com.ritense.valtimo.contract.exception.ValtimoRuntimeException;
 import com.ritense.valtimo.domain.contexts.Context;
 import com.ritense.valtimo.domain.contexts.UserContext;
-import com.ritense.valtimo.repository.ContextRepository;
-import com.ritense.valtimo.repository.UserContextRepository;
+import com.ritense.valtimo.context.repository.ContextRepository;
+import com.ritense.valtimo.context.repository.UserContextRepository;
 import lombok.RequiredArgsConstructor;
 import org.camunda.bpm.engine.RepositoryService;
 import org.camunda.bpm.engine.repository.ProcessDefinition;
