+++ core/src/test/java/com/ritense/valtimo/service/ProcessDataObjectServiceUnitTest.java
@@ -17,7 +17,7 @@
 package com.ritense.valtimo.service;
 
 import com.ritense.valtimo.domain.process.IProcessDataObject;
-import com.ritense.valtimo.repository.ProcessDataObjectRelationRepository;
+import com.ritense.valtimo.processdataobject.repository.ProcessDataObjectRelationRepository;
 import org.apache.ibatis.session.SqlSession;
 import org.apache.ibatis.session.SqlSessionFactory;
 import org.junit.jupiter.api.Assertions;
